Change type wallet amount

+++ app/src/main/java/com/kvad/totalizator/data/requestmodels/Wallet.kt
@@ -3,5 +3,5 @@ package com.kvad.totalizator.data.requestmodels
 import com.google.gson.annotations.SerializedName
 
 data class Wallet(
-    @SerializedName("amount") val amount: Int
+    @SerializedName("amount") val amount: Double
 )
